refactor: fold the software rasterizer into @runeprofile/model-renderer
One package, two entrypoints: "." stays the three.js loader the site
uses, "./rasterizer" is the zero-dependency software renderer for
Workers. They share almost no code but every exporter convention — the
format sniff, the glTF unit scale, the coordinate story — so those now
live once in src/format.ts, imported by both halves. A change to the
loader's contract lands next to the shadow implementation that must
mirror it.

Subpath isolation verified: the API worker bundle contains the
rasterizer and no trace of three.
