+++ apps/api/package.json
@@ -23,7 +23,7 @@
     "@hono/zod-openapi": "^0.18.4",
     "@hono/zod-validator": "^0.4.3",
     "@runeprofile/db": "workspace:*",
-    "@runeprofile/model-rasterizer": "workspace:*",
+    "@runeprofile/model-renderer": "workspace:*",
     "@runeprofile/runescape": "workspace:*",
     "@scalar/client-side-rendering": "0.1.6",
     "@scalar/hono-api-reference": "^0.10.13",
