+++ packages/model-renderer/src/rasterizer/index.ts
@@ -1,3 +1,4 @@
+import { detectModelFormat } from "../format";
 import { parseGlb } from "./glb";
 import { parsePly } from "./ply";
 import type { ParsedModel } from "./ply";
@@ -5,6 +6,7 @@ import { renderScene } from "./rasterizer";
 import type { ModelInstance, RenderOptions } from "./rasterizer";
 import { encodePng } from "./png";
 
+export { type ModelFormat, detectModelFormat } from "../format";
 export { parseGlb } from "./glb";
 export { parsePly } from "./ply";
 export type { ParsedModel } from "./ply";
@@ -12,25 +14,13 @@ export { renderScene } from "./rasterizer";
 export type { ModelInstance, RenderOptions } from "./rasterizer";
 export { encodePng } from "./png";
 
-/** "glTF" — the first four bytes of every binary glTF file. */
-const GLB_MAGIC = 0x46546c67;
-
 /**
  * Parses a plugin-exported model in whichever format the profile holds:
  * GLB is what the plugin uploads now, PLY is whatever a profile last
- * synced before the switch. Decided by the leading bytes, same as the
- * site's loader.
+ * synced before the switch. Uses the same sniffing rule as loadModel.
  */
 export function parseModel(data: ArrayBuffer | Uint8Array): ParsedModel {
-  const bytes = data instanceof Uint8Array ? data : new Uint8Array(data);
-  if (bytes.length >= 4) {
-    const magic =
-      bytes[0]! | (bytes[1]! << 8) | (bytes[2]! << 16) | (bytes[3]! << 24);
-    if ((magic >>> 0) === GLB_MAGIC) {
-      return parseGlb(bytes);
-    }
-  }
-  return parsePly(bytes);
+  return detectModelFormat(data) === "glb" ? parseGlb(data) : parsePly(data);
 }
 
 /** Renders a single model (GLB or PLY) straight to a PNG. */
