+++ packages/model-renderer/src/rasterizer/ply.ts
