+++ packages/model-renderer/src/rasterizer/png.ts
