+++ packages/model-renderer/src/rasterizer/rasterizer.ts
